fixed udp id

+++ Server/server.hpp
@@ -15,11 +15,10 @@ private:
 	sf::SocketSelector selector;
 	sf::UdpSocket udp;
 
-	std::map<sf::Uint16, Client> clients;
+	std::map<sf::Uint32, Client> clients;
 	std::vector<sf::TcpSocket*> uninitialized;
 
-	std::set<sf::Uint16> publicLobby;
-
-	sf::Uint16 currentId = 0;
+	std::set<sf::Uint32> publicLobby;
+	sf::Uint32 currentId = 0;
 };
